# services/external/api-gateway/api/server/api.yaml
openapi: 3.0.3
info:
  title: codex-k8s api-gateway
  version: 0.2.0
  description: |
    External webhook + staff API contract for codex-k8s.
servers:
  - url: https://platform.codex-k8s.dev
paths:
  /api/v1/webhooks/github:
    post:
      summary: Ingest GitHub webhook delivery
      operationId: ingestGithubWebhook
      tags: [webhooks]
      parameters:
        - in: header
          name: X-GitHub-Event
          required: true
          schema:
            type: string
        - in: header
          name: X-GitHub-Delivery
          required: true
          schema:
            type: string
        - in: header
          name: X-Hub-Signature-256
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Accepted and enqueued for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestGitHubWebhookResponse"
        "200":
          description: Duplicate delivery already ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestGitHubWebhookResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Internal:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_argument
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - failed_precondition
            - resource_exhausted
            - deadline_exceeded
            - unavailable
            - canceled
            - internal
        message:
          type: string
        field:
          type: string
    IngestGitHubWebhookResponse:
      type: object
      additionalProperties: false
      required: [correlation_id, status, duplicate]
      properties:
        correlation_id:
          type: string
        run_id:
          type: string
        status:
          type: string
          enum: [accepted, duplicate, ignored]
        duplicate:
          type: boolean
...